pyproject.toml
[tool.poetry]
name = "transcriber"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "transcriber"}]

include = [
    { path = "transcriber/models/*.yaml", format = ["sdist", "wheel"] },
    { path = "transcriber/models/*.bin", format = ["sdist", "wheel"] },
]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.12"
torch = { version = "2.5.0", source = "pytorch" }
openai-whisper = "^20240930"
pyannote-audio = "^3.3.2"
pydub = "^0.25.1"
ffmpeg = "^1.4"
pyyaml = "^6.0.2"
pydantic = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=transcriber --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
